Add in roomservice.
This is a new implementation of roomservice, use the new Github search API
to reduce the memory footprint, and make it less processor intensive allow
for unofficial devices to fetch dependencies. just specify your Github
organization.

To fetch dependencies:
    add a file called 'omni.dependencies' to the root of your device tree
    an example configuration for the dependency file is:

[
  {
    "repository": "android_device_sony_common",
    "target_path": "device/sony/common"
  },
  {
    "repository": "legacybop/android_hardware_qcom_display-legacy",
    "target_path": "hardware/qcom/display-legacy",
    "revision": "omni-4.3"
  }

Please note that comments can not go into this file. It must be valid JSON.

Change-Id: Icf6e107a1f52b27f5148090ccbc1e6df1077f77c