Allow building DTBO image.
Taimen uses the dtbo.img as an DTB Overlay in order to
use the same defconfig for both, taimen and wahoo.

The DTBO image is built with "m dtbo.img". Allow automatically
building it when setting TARGET_NEEDS_DTBOIMAGE to true.

Thanks to luca020400 for help with the kernel.mk logic.

[raymanfx] Add AVB hash footer support.

Change-Id: Iaddafaa98e3e84880462a2c35b899e3fea21e161